Refactor: 로그인/회원가입 부분 리팩토링(컨테이너화)
